requirements.txt
Flask
Flask-SQLAlchemy
Flask-WTF
Flask-Login
Flask-Mail
PyMySQL
python-dotenv
bcrypt
python-dateutil
qrcode
Pillow
itsdangerous
email-validator